profiles/laptop: ensure primary battery is used
Some laptop (eg last ThinkPad before they dropped removable batteries)
have a two battery system and the udev event will act when the first
battery dies, which usually is less than 20-30% of your entire battery
life, so not great so great.

This changes the default behaviour to ensure this only happens on BAT0
which, hopefully, should always be the primary battery. Other
assumptions in navi will break if this isn't the case, so I'll have
other things to worry about in this case